fix pressing enter on open team. move things that aren't rows out of rows
