+++ shared/chat/inbox/filter-row/container.tsx
@@ -1,7 +1,7 @@
-import * as Chat2Gen from '../../../../actions/chat2-gen'
-import * as RouteTreeGen from '../../../../actions/route-tree-gen'
-import {appendNewChatBuilder} from '../../../../actions/typed-routes'
-import * as Container from '../../../../util/container'
+import * as Chat2Gen from '../../../actions/chat2-gen'
+import * as RouteTreeGen from '../../../actions/route-tree-gen'
+import {appendNewChatBuilder} from '../../../actions/typed-routes'
+import * as Container from '../../../util/container'
 import ConversationFilterInput from '.'
 
 type OwnProps = {
