+++ shared/chat/inbox/filter-row/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
-import * as Kb from '../../../../common-adapters'
-import * as Styles from '../../../../styles'
-import * as Platforms from '../../../../constants/platform'
+import * as Kb from '../../../common-adapters'
+import * as Styles from '../../../styles'
+import * as Platforms from '../../../constants/platform'
 
 export type Props = {
   appendNewChatBuilder: () => void
