+++ shared/chat/inbox/index.native.tsx
@@ -6,7 +6,7 @@ import * as T from './index.d'
 import * as Types from '../../constants/types/chat2'
 import BigTeamsDivider from './row/big-teams-divider/container'
 import BuildTeam from './row/build-team'
-import ChatInboxHeader from './row/chat-inbox-header/container'
+import ChatInboxHeader from './header/container'
 import InboxSearch from '../inbox-search/container'
 import TeamsDivider from './row/teams-divider/container'
 import UnreadShortcut from './unread-shortcut'
