+++ shared/constants/chat2/index.tsx
@@ -20,6 +20,7 @@ import HiddenString from '../../util/hidden-string'
 import {getFullname} from '../users'
 import {memoize} from '../../util/memoize'
 import * as TeamConstants from '../teams'
+import flags from '../../util/feature-flags'
 
 export const defaultTopReacjis = [':+1:', ':-1:', ':tada:', ':joy:', ':sunglasses:']
 const defaultSkinTone = 1
@@ -162,8 +163,14 @@ export const isCancelledAudioRecording = (audioRecording: Types.AudioRecordingIn
 }
 
 export const getInboxSearchSelected = (inboxSearch: Types.InboxSearchInfo) => {
-  if (inboxSearch.selectedIndex < inboxSearch.nameResults.length) {
-    const maybeNameResults = inboxSearch.nameResults[inboxSearch.selectedIndex]
+  const {selectedIndex, nameResults, openTeamsResults, textResults} = inboxSearch
+  const firstTextResultIdx = flags.openTeamSearch
+    ? openTeamsResults.length + nameResults.length
+    : nameResults.length
+  const firstOpenTeamResultIdx = nameResults.length
+
+  if (selectedIndex < firstOpenTeamResultIdx) {
+    const maybeNameResults = nameResults[selectedIndex]
     const conversationIDKey =
       maybeNameResults === null || maybeNameResults === undefined
         ? undefined
@@ -174,8 +181,10 @@ export const getInboxSearchSelected = (inboxSearch: Types.InboxSearchInfo) => {
         query: undefined,
       }
     }
-  } else if (inboxSearch.selectedIndex < inboxSearch.nameResults.length + inboxSearch.textResults.length) {
-    const result = inboxSearch.textResults[inboxSearch.selectedIndex - inboxSearch.nameResults.length]
+  } else if (flags.openTeamSearch && selectedIndex < firstTextResultIdx) {
+    return null
+  } else if (selectedIndex >= firstTextResultIdx) {
+    const result = textResults[selectedIndex - firstTextResultIdx]
     if (result) {
       return {
         conversationIDKey: result.conversationIDKey,
@@ -183,6 +192,7 @@ export const getInboxSearchSelected = (inboxSearch: Types.InboxSearchInfo) => {
       }
     }
   }
+
   return null
 }
